more work on job server, split up into more files.
