+++ Sources/HerbieHelper/JobServer/MessageTypes/Sample/HerbieRequest.swift
@@ -1,0 +1,14 @@
+import Foundation
+
+struct HerbieRequest: Codable {
+    let formula: String
+    let seed: Int
+
+    public var json: String {
+        let encoder = JSONEncoder()
+        // encoder.outputFormatting = .prettyPrinted
+        let data = try! encoder.encode(self)
+        let jsonString = String(data: data, encoding: .utf8)!
+        return jsonString
+    }
+}
